修改参数类型

+++ ssm-demo/src/main/java/com/jusfoun/mapper/ds0/SysGovMapper.java
@@ -16,7 +16,7 @@ public interface SysGovMapper extends BaseWithAssociateMapper<SysGov> {
 	 *            ID
 	 * @return 名称
 	 */
-	String selectNameByPrimaryKey(Integer id);
+	String selectNameByPrimaryKey(Long id);
 
 	/**
 	 * 描述 : 根据父节点ID查询子节点列表. <br>
@@ -27,5 +27,5 @@ public interface SysGovMapper extends BaseWithAssociateMapper<SysGov> {
 	 *            父节点ID
 	 * @return 子节点集合
 	 */
-	List<SysUser> selectByParentId(Integer parentId);
+	List<SysUser> selectByParentId(Long parentId);
 }
